// core/connectionManager/types.ts
export enum COMMUNICATION_MANAGER_EVENTS {
  ADD_PARTICIPANT = "ADD_PARTICIPANT",
  REMOVE_PARTICIPANT = "REMOVE_PARTICIPANT",
}

export enum INCOMING_SOCKET_EVENTS {
  ROOM_CREATED = "ROOM_CREATED",
  ROOM_JOINED = "ROOM_JOINED",
  NEW_GUEST_JOINED = "NEW_GUEST_JOINED",
  OFFER = "OFFER",
  ANSWER = "ANSWER",
  BROADCAST = "BROADCAST",
  GUEST_LEFT = "GUEST_LEFT",
  ADD_ICE_CANDIDATE = "ADD_ICE_CANDIDATE",
}

export enum OUTGOING_SOCKET_EVENTS {
  JOIN = "JOIN",
  SEND_OFFER = "SEND_OFFER",
  SEND_ANSWER = "SEND_ANSWER",
  LEAVE = "LEAVE",
  SEND_ICE_CANDIDATE = "SEND_ICE_CANDIDATE",
}

export type SocketEvent = INCOMING_SOCKET_EVENTS | OUTGOING_SOCKET_EVENTS;

export enum WEB_RTC_EVENTS {
  TRACK = "track",
  ICE_CANDIDATE = "icecandidate",
  ICE_CANDIDATE_ERROR = "icecandidateerror",
  ICE_CONNECTION_STATE_CHANGE = "iceconnectionstatechange",
  DATA_CHANNEL = "dataChannel",
  CONNECTION_STATE_CHANGE = "connectionstatechange",
  ICE_GATHERINGS_STATE_CHANGE = "icegatheringstatechange",
  NEGOTIATION_NEEDED = "negotiationneeded",
  SIGNALLING_STATE_CHANGE = "signalingstatechange",
}